Create Installation_with_R_v1.qmd
Quarto document with the full installation of SpectriPy using only R and reticulate. Explain step by step how to install python with miniconda, to then install all python packages, and all R packages that would be needed.

An example to test the installation is present at the end of the document.